+++ lib/khmer_intl.dart
@@ -7,7 +7,7 @@ export 'src/date/khmer_date_format.dart';
 export 'src/date/khmer_date_presets.dart';
 export 'src/time/khmer_relative_time.dart';
 export 'src/plural/khmer_plural.dart';
-export 'src/lunar/khmer_lunar_calendar.dart';
+export 'src/lunar/khmer_chhankitek.dart';
 
 export 'src/extensions/int_extensions.dart';
 export 'src/extensions/num_extensions.dart';